+++ dbdaora/hash/_tests/datastore/test_integration_service_hash_aioredis_datastore_get_many.py
@@ -15,12 +15,16 @@ async def test_should_get_many(
     fake_entity2,
 ):
     await fake_service.repository.memory_data_source.hmset(
-        'fake:fake', *itertools.chain(*serialized_fake_entity.items())
+        'fake:other_fake:fake',
+        *itertools.chain(*serialized_fake_entity.items()),
     )
     await fake_service.repository.memory_data_source.hmset(
-        'fake:fake2', *itertools.chain(*serialized_fake_entity2.items())
+        'fake:other_fake:fake2',
+        *itertools.chain(*serialized_fake_entity2.items()),
     )
-    entities = await fake_service.get_many('fake', 'fake2')
+    entities = await fake_service.get_many(
+        'fake', 'fake2', other_id='other_fake'
+    )
 
     assert entities == [fake_entity, fake_entity2]
 
@@ -34,7 +38,9 @@ async def test_should_get_many_from_cache(
     )
     fake_service.cache['fake'] = fake_entity
     fake_service.cache['fake2'] = fake_entity2
-    entities = await fake_service.get_many('fake', 'fake2')
+    entities = await fake_service.get_many(
+        'fake', 'fake2', other_id='other_fake'
+    )
 
     assert entities == [fake_entity, fake_entity2]
     assert not fake_service.repository.memory_data_source.hgetall.called
@@ -49,18 +55,22 @@ async def test_should_get_many_from_fallback_after_open_circuit_breaker(
     fake_pipeline.return_value.execute = asynctest.CoroutineMock(
         side_effect=RedisError
     )
-    key = fake_service.repository.fallback_data_source.make_key('fake', 'fake')
+    key = fake_service.repository.fallback_data_source.make_key(
+        'fake', 'other_fake', 'fake'
+    )
     await fake_service.repository.fallback_data_source.put(
         key, dataclasses.asdict(fake_entity)
     )
     key = fake_service.repository.fallback_data_source.make_key(
-        'fake', 'fake2'
+        'fake', 'other_fake', 'fake2'
     )
     await fake_service.repository.fallback_data_source.put(
         key, dataclasses.asdict(fake_entity2)
     )
 
-    entities = await fake_service.get_many('fake', 'fake2')
+    entities = await fake_service.get_many(
+        'fake', 'fake2', other_id='other_fake'
+    )
 
     assert entities == [fake_entity, fake_entity2]
     assert fake_service.logger.warning.call_count == 1
@@ -81,7 +91,10 @@ async def test_should_get_many_with_fields(
         'fake:fake2', *itertools.chain(*serialized_fake_entity2.items())
     )
     entities = await fake_service.get_many(
-        'fake', 'fake2', fields=['id', 'integer', 'inner_entities']
+        'fake',
+        'fake2',
+        fields=['id', 'other_id', 'integer', 'inner_entities'],
+        other_id='other_fake',
     )
     fake_entity.number = None
     fake_entity.boolean = False
@@ -101,7 +114,10 @@ async def test_should_get_many_from_cache_with_fields(
     fake_service.cache['fake'] = fake_entity
     fake_service.cache['fake2'] = fake_entity2
     entities = await fake_service.get_many(
-        'fake', 'fake2', fields=['id', 'integer', 'inner_entities']
+        'fake',
+        'fake2',
+        fields=['id', 'other_id', 'integer', 'inner_entities'],
+        other_id='other_fake',
     )
     fake_entity.number = None
     fake_entity.boolean = False
@@ -137,7 +153,10 @@ async def test_should_get_many_from_fallback_after_open_circuit_breaker_with_fie
     fake_entity2.boolean = False
 
     entities = await fake_service.get_many(
-        'fake', 'fake2', fields=['id', 'integer', 'inner_entities']
+        'fake',
+        'fake2',
+        fields=['id', 'other_id', 'integer', 'inner_entities'],
+        other_id='other_fake',
     )
 
     assert entities == [fake_entity, fake_entity2]