+++ dbdaora/hash/_tests/datastore/test_integration_service_hash_aioredis_datastore_get_one.py
@@ -13,7 +13,7 @@ async def test_should_get_one(
     await fake_service.repository.memory_data_source.hmset(
         'fake:fake', *itertools.chain(*serialized_fake_entity.items())
     )
-    entity = await fake_service.get_one('fake')
+    entity = await fake_service.get_one('fake', other_id='other_fake')
 
     assert entity == fake_entity
 
@@ -23,13 +23,16 @@ async def test_should_get_one_with_fields(
     fake_service, serialized_fake_entity, fake_entity
 ):
     await fake_service.repository.memory_data_source.hmset(
-        'fake:fake', *itertools.chain(*serialized_fake_entity.items())
+        'fake:other_fake:fake',
+        *itertools.chain(*serialized_fake_entity.items()),
     )
     fake_entity.number = None
     fake_entity.boolean = False
 
     entity = await fake_service.get_one(
-        'fake', fields=['id', 'integer', 'inner_entities']
+        'fake',
+        fields=['id', 'other_id', 'integer', 'inner_entities'],
+        other_id='other_fake',
     )
 
     assert entity == fake_entity
@@ -43,7 +46,7 @@ async def test_should_get_one_from_cache(
         asynctest.CoroutineMock()
     )
     fake_service.cache['fake'] = fake_entity
-    entity = await fake_service.get_one('fake')
+    entity = await fake_service.get_one('fake', other_id='other_fake')
 
     assert entity == fake_entity
     assert not fake_service.repository.memory_data_source.hgetall.called
@@ -56,12 +59,14 @@ async def test_should_get_one_from_fallback_after_open_circuit_breaker(
     fake_service.repository.memory_data_source.hgetall = asynctest.CoroutineMock(
         side_effect=RedisError
     )
-    key = fake_service.repository.fallback_data_source.make_key('fake', 'fake')
+    key = fake_service.repository.fallback_data_source.make_key(
+        'fake', 'other_fake', 'fake'
+    )
     await fake_service.repository.fallback_data_source.put(
         key, dataclasses.asdict(fake_entity)
     )
 
-    entity = await fake_service.get_one('fake')
+    entity = await fake_service.get_one('fake', other_id='other_fake')
 
     assert entity == fake_entity
     assert fake_service.logger.warning.call_count == 1