+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix DatastoreDataSource.make_key method
+
 ### 0.2.0 - 2020-04-24
 
  - Create dbdaora package